move test file out of tests dir in flags module (#293)
