+++ flags/test.ts
@@ -1,14 +1,14 @@
 // Copyright 2018-2019 the Deno authors. All rights reserved. MIT license.
-import "./tests/all_bool.ts";
-import "./tests/bool.ts";
-import "./tests/dash.ts";
-import "./tests/default_bool.ts";
-import "./tests/dotted.ts";
-import "./tests/kv_short.ts";
-import "./tests/long.ts";
-import "./tests/num.ts";
-import "./tests/parse.ts";
-import "./tests/short.ts";
-import "./tests/stop_early.ts";
-import "./tests/unknown.ts";
-import "./tests/whitespace.ts";
+import "./all_bool_test.ts";
+import "./bool_test.ts";
+import "./dash_test.ts";
+import "./default_bool_test.ts";
+import "./dotted_test.ts";
+import "./kv_short_test.ts";
+import "./long_test.ts";
+import "./num_test.ts";
+import "./parse_test.ts";
+import "./short_test.ts";
+import "./stop_early_test.ts";
+import "./unknown_test.ts";
+import "./whitespace_test.ts";
